Add basic tests for PATH and QUERY operation parameters
- Integrate basic serialization tests for operation parameters.
Add testing of QString, QList<int>, QJsonValue, OAIObject types
for all PATH and QUERY combinations supported in 3.1.1.
To check covered variations, see the table Basic-types-testcases.xlsx
attached in the task bellow.
To check serialization rules, see the doc used for implementation
of OP here:
https://spec.openapis.org/oas/v3.1.1.html#style-values

NOTE: All non-passed test-cases are being skipped with QSKIP macro.
QEXPECTED_FAIL macro is not so convenient here, because it requires
a lot of if/else for data-driven test-cases.

NOTE: The fix of all test-cases are done in the next commit.

Task-number: QTBUG-138564
Change-Id: I79d0ffe60ebee06fbd4c0dc583eef4477bac38c4
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>