Fix wildcard disconnect() in callback code
Commit qtbase/a79d5b8d0856a8f6c4de56a8d982c84ea20905a6 added a
warning when a wildcard disconnect() call disconnects from the
QObject::destroyed() signal, because that signal can be relevant
for Qt internally.

We use the wildcard disconnect() in every operation callback to
avoid receiving more signals from the rest reply that is handled,
meaning that every operation call was generating the new warning.

Strictly speaking, this disconnect() should not be required at
all, but this commit plays safe and replaces the wildcard call to

  netReply->disconnect(this);

which disconnects the reply only from this specific object, but
not from the QNAM and/or other internal instances that can be
responsible for managing the reply.

Fixes: QTBUG-138857
Change-Id: I291de154d7a1d067c48710e07fb6b5dc16fcc2c3
Reviewed-by: Tatiana Borisova <tatiana.borisova@qt.io>
Reviewed-by: Juha Vuolle <juha.vuolle@qt.io>