Add fromByteArray() helpers for primitive types
... and use them in the operation callback code, where
constructing an intermediate QString really seems unnecessary for
most of the possible types.

In some cases (e.g. date and time classes) there's no optimal path
for QByteArray input, so the implementation actually falls back
to the QString version.

This patch does not touch the enum/object-model hierarchy, so
their implementation also unconditionally converts the input
to QString. It requires some more investigation to see if it can
be changed or not.

It seems that most of the fromStringValue() helpers are now unused
or only used as implementation detail of the newly introduced
fromByteArray() methods. However, let's keep them for now, because
they might be needed later when improving the response handling.

Fixes: QTBUG-137429
Change-Id: Ibcbe47c1e6f7e06e181056adadb98f50e01d2cfa
Reviewed-by: Tatiana Borisova <tatiana.borisova@qt.io>