Add warning when style value is invalid for path params
Add warning during generation in CppQt6AbstractCodegen.java and in
the generated code in api-body.mustache.
Inform the user that the provided style value will be ignored, and the
default 'simple' style will be used instead.
Set style value to 'simple' in the java code in case of invalid style.

Add a new path operation with invalid style. Add test.

Task-number: QTBUG-137408
Change-Id: I885d4b0838631f60600d1334628068b7fb29321d
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>