Use Qt::StringLiterals in HttpRequest
Also use u'' to create QChar from char16_t characters.

This patch does not touch QNetworkRequest::setHeader() calls, because
it takes QByteArray and QVariant as its input parameters.
Similarly, do not touch the QUrl::toPercentEncoding().replace() call,
because if operates on a QByteArray, so the arguments are also byte
arrays.

As a drive-by,  make some variables in the modified lines const.

Task-number: QTBUG-137390
Change-Id: I5cb68cbf778d81d296bb6f6c69f6de84bf8f7241
Reviewed-by: Tatiana Borisova <tatiana.borisova@qt.io>