Introduce a new serializeMapValue template function
- Rename toStringValue(const QMap<QString, T> &val ....)
to serializeMapValue(const QMap<QString, T> &mapValue ...)
and move it to api-base-header class.
The reason is toStringValue() should not have additional arguments,
like keyValueDelimiter, itemDelimiter.
At the same moment those arguments are OK for serializeArrayValue() and
serializeJsonValue().
Also, it is better to edit API of serializeArrayValue(),
serializeJsonValue(), serializeMapValue() together. See the next patches.

- Delete QString toStringValue(const QList<T> &val, QAnyStringView
delimiter), since toStringValue() should not have additional
arguments. And this method is not used after this change.

- Delete duplicating logic:
parameterToString() and optionParameterToString(), since
  - parameterToString() was not used anywhere.
  - optionParameterToString() is replaced with other function calls and
also not used anymore.

Change-Id: Ib354f840022d0a19e2d32b07972a09f604edf5f2
Signed-off-by: Tatiana Borisova <tatiana.borisova@qt.io>
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>