Fix deepobject serialization for anytype object parameters
{{prefix}}BaseApi::serializeJsonValue() works well for all types,
including QJsonValue::Object, except when the style is deepObject.
Previously, The parameter name was not appeneded with [] to the path for
anytype objects.

Remove the deepobject logic for object query parameters ({{#isModel}})
because this is now handled inside serializeJsonValue().

For most styles, the parameter name is included in the suffix.
However, the deepObejct suffix was empty. Store the parameter name in it.

Change-Id: I211d333f4d5af396d6b0553571084b02ddfd8e16
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>