Introduce SerializationOptions struct
The helper functions for serialization took too many parameters,
most of them being either QString or bool.
As a result, it was very easy to pass the arguments in the wrong
order.

This patch introduces a new enum for boolean options, and a new
struct to hold all serialization options. The struct is now
populated in a single place and then passed to all serialization
methods.

Task-number: QTBUG-140222
Change-Id: I64690202979a0206fea9ea1d05c9e345168869ce
Reviewed-by: Tatiana Borisova <tatiana.borisova@qt.io>