Add go server for petstore tests
For now, as a separate application for manual testing.
This commit does not touch the old Qt5-based server.

So far, the following changes to the test code are required:
* gin parser does not understand DateTime string without timezone,
  so we have to use QDateTime::currentDateTimeUtc(). This also
  works fine with Qt5 server
* the PetApiTests::uploadPetFileTest() test was providing an
  incorrect name for the file that it tried to upload. As a
  result, the file contents was never actually sent. This was
  unnoticed, because the C++ server did not really try to read
  the received file. Fixed the test to provide a proper filename.

Note that all scripts still use the old Qt5-based server.
There will be a separate follow-up commit to remove the old
server and use the new server.

Task-number: QTBUG-137879
Change-Id: Idd759743e7cf5da54af5cc4055122e94c143ab8b
Reviewed-by: Tatiana Borisova <tatiana.borisova@qt.io>