Delete AUTORCC and AUTOUIC from a Client library generation
- the usage is not required, because the Qt6 Client library doesn't
have any *.ui files or resource files.
Moreover, there is no plans at all to generate such files
for the Client library.

- also, the usage of AUTORCC and AUTOUIC in a Qt6 Client library
leads to a configuration error, if build the qtopenapi
as a sub-module of qt5/* project.

So we should delete set(CMAKE_AUTOUIC ON) and set(CMAKE_AUTORCC ON)
from CMakeLists.txt

Task-number: QTBUG-140130
Change-Id: Id35ee62aea5eaabc2c53a6d21ff564b1bc3fd530
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>