Fix OperationParameters::severalHeaderParameters() test
The test was missing the QTRY_COMPARE_EQ in the end.
It's not only a logical error, but it was also leading to the fact
that the `done` variable went out of scope by the time the callback
was invoked, so we had an AddressSanitizer stack-use-after-scope
issue.

Adding QTRY_COMPARE_EQ both fixes the test logic and prevents the
variable to get out of scope.

Change-Id: Ifd2b35abbe9b468b54da222219babd213795fff6
Reviewed-by: Rym Bouabid <rym.bouabid@qt.io>
Reviewed-by: Tatiana Borisova <tatiana.borisova@qt.io>