Fix zlib usage when compression is enabled
Introduce the following changes to the CMakeLists.txt template
if compression is enabled:
- use find_package(ZLIB REQUIRED) to properly detect the library
- add ZLIB_INCLUDE_DIRS to the list of include directories in order
  to find the zlib.h include
- use ZLIB::ZLIB as a private link target instead of ZLIB_LIBRARIES,
  mainly because that's the recommended way mentioned in the CMake
  docs.

Create a new "compression" test and update the generate.sh script
to enable compression for it.
For now the test does not have any server because it simply does
a local compress-decompress roundtrip test. It also serves as a
compile-time check that zlib is properly found and linked.

As a drive-by: fix formatting of the {{contentCompression}} blocks
to make them more readable.

Task-number: QTBUG-139963
Change-Id: I525592246e831d9038a090d6b7fa125ac23086a7
Reviewed-by: Tatiana Borisova <tatiana.borisova@qt.io>