Add export macros to enable builds on Windows
As we build the generated OpenAPI clients as libraries, we
actually need to export symbols. That was not noticed while we
only had builds on Linux and macOS, because AFAIK all symbols are
exported there by default.

Now when we want to add Windows builds, we actually need to
introduce the proper export macros, that would expand to dllexport
or dllimport.
We use the logic from Qt, also relying on the QT_SHARED and
QT_STATIC definitions. That should be fine since we're anyway
using qt_add_library() to create the libraries, and it inherits
the build type from Qt.

Since the client is separated into a common library and a client
library, we need to introduce two separate export macros, because
otherwise we won't be able to link the client library with the
common library.
We add two new templates:
* common/common-exports.mustache for the exports from the common
  library
* exports.mustache for the exports from the client library

Task-number: QTBUG-138968
Change-Id: Ib6464cb542acd7783bfcb2dbbcd565c322128173
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>
Reviewed-by: Tatiana Borisova <tatiana.borisova@qt.io>