Temporary build OpenAPI library
QtOpenAPI does not have any Qt-specific code, it only generates
client libraries from provided templates.
Because of that, the module was built as an interface library.
This, however, causes bugs in release-and-debug configuration.
Specifically, an attempt to configure tests does nothing because
the QT_FEATURE_openapi_generator feature is not found, and so all
tests are skipped.

Ideally this should be fixed at the cmake level in qtbase, but as
a temporary solution simply add a dummy cpp file with a dummy
exported function.
Also add missing licenses to LICENSES subdir.

Task-number: QTBUG-138968
Change-Id: I8641cf394bf1d3e05ed4c6e88e3f8554a4548ce5
Reviewed-by: Rym Bouabid <rym.bouabid@qt.io>
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>