Fix mediatype test on Windows
The test sends the contents of the text file that ends with a
newline to the server, and then reads the same data back.
Because of the fact that Windows and Unix use different line
endings, it is not possible to reliably test the data across
these systems.
To fix it, use {QString,QByteArray}::trimmed() in the comparison,
thus discarding the newline combinations.

Task-number: QTBUG-138968
Change-Id: If7a8c45705912c06286bec2facf079ec17b61804
Reviewed-by: Rym Bouabid <rym.bouabid@qt.io>
Reviewed-by: Tatiana Borisova <tatiana.borisova@qt.io>