Adapt tests to Windows CI
The tests link to client libraries, which in turn link to their
respective common libraries.

On windows, when we run ctest, it is already aware of the Qt
installation path, but it fails to find the client and common libs.
Fix it by explicitly copying the libraries to the same directory
where the test executable is located.

Task-number: QTBUG-138968
Change-Id: I34c510762705c5c03fe7c0dcdc59d9d710aef250
Reviewed-by: Tatiana Borisova <tatiana.borisova@qt.io>