Make Enum base class pure virtual
The implementation of the base class is not used anywhere.
The model that are derived from Enum provide their own
implementations of all virtual methods based on the members that
they generate. None of these methods ever calls the base
implementation.

This commit makes the base class pure virtual and also add
(in)equality operators and qHash() overload as hidden friends.

Fixes: QTBUG-139919
Change-Id: I646c143acae8fb7add1e97f4b9ea0ae08ddee257
Reviewed-by: Tatiana Borisova <tatiana.borisova@qt.io>