Make Object base class pure virtual
Similarly to the Enum, the implementation of the base class
is not used anywhere.
The models that are derived from Object provide their own
implementations of all virtual methods based on the members that
they generate. None of these methods ever calls the base
implementation.

This commit makes the base class pure virtual and also converts
the (in)equality operators and qHash() overloads to hidden friends
instead of having them in the global namespace. This makes sure
that they are only found by ADL lookup.

Change-Id: Icc21ae332c37115bf7eab1699c2a7541b55956f6
Reviewed-by: Tatiana Borisova <tatiana.borisova@qt.io>