Use std::string to store SerializationOptions::style
Style can only be one of the pre-defined options, all of which are
ASCII-only. As a result, storing it in a QString is an overhead.

We could use QByteArray, but using this class to store strings is
usually a bad design choice, because, as the name implies, it
should be used for storing raw bytes.
So, use std::string instead.

Update the functions that take the style as a parameter to take
std::string_view, as they are only inspecting it.

Using std::string instead of QString also allows us to drop all the
usages of Qt string literal ""_L1 when comparing the value.

Theoretically we could do the same optimization for the assignOperator
and delimiter members of the SerializationOptions struct, but those
are later used to construct the resulting QString, so would be
converted to QString anyway.

Task-number: QTBUG-137432
Change-Id: I812c460762541d6f992f73b9d46f735c495a2dcd
Reviewed-by: Tatiana Borisova <tatiana.borisova@qt.io>