CMake: Improve build system dependency handling and summary
Rewrite configure.cmake to check for the required 3rd party
dependencies in a more precise manner.

Use find_package(Java) to find the java compiler.

Add detailed configure summary information for the various
dependencies.

Improve the wording of various diagnostic messages.

Replace the early returns in src/tools/qtopenapi-generator
with a check for the appropriate feature in the parent CMakeLists.txt.

Change-Id: I079b240fe46e49e1d1232d356451438e15c04a25
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>