CMake: Add WrapOpenAPIGenerator and Java dependency for OpenApi package
When a user project calls find_package(Qt6OpenapiPrivate), to be able
to use the Qt provided generator plugin, we also need to find the
upstream generator CLI and the java runtime

Add a dependency to find the WrapOpenAPIGenerator and WrapOpenAPIJava
packages when looking for Qt6OpenApiPrivate.

Remove the manual dependency in the tests subdir, it will be handled
automatically now.

Change-Id: Idcd0fbac558972909f1fa7c55e4ff1da73e2d61f
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>