CMake: Rename the OpenapiPrivate module to OpenApi
User projects need to find a Qt package to be able to use the public
cmake API. Which means the module can't have a Private suffix.

Rename the module to OpenApi (as opposed to Openapi, note the casing)
and rename some of the supporting files to use the same casing.

The module was previously a real cpp library because we don't export
configure features for INTERFACE libraries properly in qtbase, and the
openapi feature was checked in the tests subdirectory.

Change the module type back to HEADER_MODULE (INTERFACE_LIBRARY),
remove the dummy.cpp source file, add a bunch of options to generate
as few things as possible.

Change the tests to check for the existence of the
Qt6::QtOpenAPIGeneratorJar and
WrapOpenAPIGenerator::WrapOpenAPIGenerator targets instead of the
openapi_generator feature to decide whether all dependencies are
available.

Ideally we wouldn't have to create a module at all, and place the
imported generator target in a find_package(Qt6OpenApiTools)
package, but that's currently not possible to do without changes in
qtbase.
This is the best we can do at the moment.

Change-Id: I4d7561cf6d14eb554e482ebb5091f1b41a740f9a
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>