CMake: Move feature evaluation to a non-module early subdirectory
Previously, if some build dependency was not found, and the generator
plugin was not built, we still ended up installing the Qt OpenAPI
cmake package with the macros and the extra targets file.

This is undesirable, as the package should not be installed at all if
the generator is not built.

Move the evaluation of the features to be done earlier before the
openapi subdirectory is added. Then add the openapi subdirectory
only if the generator feature is enabled.

Use the NO_MODULE option of qt_feature_module_begin / end to evaluate
the features, while also showing the configure summary messages.

Change-Id: I493144b83009a3d8b5833619297931aa8aa0a91d
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>