Add a REUSE.toml file to the qtopenapi repo
- Add the REUSE.toml file to trigger a license validation properly.
The file documents copyright and licensing information
when it cannot be included in-file due to the file format constraints
or due to any other complications.
- Mention in the file test resources (like go files, png images,
yaml files, etc) under "LicenseRef-Qt-Commercial OR GPL-3.0-only".
Also include into the file *.mustache files and *.xml
java-project files under "Apache-2.0".
- Delete Qt-GPL-exception-1.0.txt file, since "reuse" tool
complains the file is not used. And it fails license validation.

Change-Id: I5429a2b256961cdecbc14ca54a8acafc3fc555a0
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>