CMake: Split OpenApi module into an additional tools package
The qt generator plugin can be considered as an arch-independent
tool that can be re-used across different platforms. Thus it makes
sense to package it in a separate OpenApiTools cmake package, rather
than as part of the main OpenApi cmake package. This follows the
precedent of keeping tools in a separate package.

Create a new OpenApiTools package that will create the generator
plugin cmake IMPORTED target, as well as expose the openapi public
CMake API for using the tool.

Make the main OpenApi package depend on the OpenApiTools package.
Make the java and upstream generator dependencies be dependencies of
the tools package rather than the main module.

Change-Id: I42f9ae13fd0bc56f2610bc081f78f4847917df5d
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>