ServerVariable: fix the general case when no enum is defined
According to the spec [0], server variables may have an enum field
which defines the allowed set of the values.
This part was implemented correctly.

However, if the enum field is not present, any value should be
accepted. This part was wrong, because the code simply used a
default value as a single element of the set. As a result, no
other values could be specified.

This patch fixes the logic and adds a test that covers this case.

[0]: https://spec.openapis.org/oas/v3.1.0.html#server-variable-object

Task-number: QTBUG-141948
Change-Id: Ib4366e07a769f27ce615eecd3a2d86b4e33fd42c
Reviewed-by: Tatiana Borisova <tatiana.borisova@qt.io>