Make WrapOpenAPIGenerator a requirement for openapi_generator
The previous assumption that the upstream generator was only
needed to build the tests is now wrong, because we need to build
the Qt::OpenApiCommon library.

Go even one step further and make the upstream generator a strict
requirement for the entire module build, because it is now not
really usable without the common library.

Amends bec8a66d52aaa32f3caf019ac1dbcf314462989b

Task-number: QTBUG-142230
Change-Id: I7572ccc9bfe5120bb560dbc92369c1242d49f54a
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>