Add BaseApi::operations() slot
The slot will return a list of all supported operations.

We have a lot of server configuration APIs that use the name of
the operation to identify the servers, but we do not have any APIs
that allow to get the list of all available operations.

This patch fixes it.

Found in API review.

Task-number: QTBUG-142269
Pick-to: 6.11
Change-Id: I3a3620006fb75fbe40845489a4ded9d18ad56243
Reviewed-by: Tatiana Borisova <tatiana.borisova@qt.io>