Separate a 'prefix' for common files and a 'modelNamePrefix'
- the prefixes were mixed up together.
But the 'modelNamePrefix' should NOT be connected somehow with common
library files. The 'modelNamePrefix' should affect only names of
a client library.

But for common files we should have a separate option. So,
I add the 'prefix' option to the 'additional-properties'
to split a prefixes.

- Rename MODEL_NAME_PREFIX to CLIENT_PREFIX, because
it is responsible for client files.
Note: it's always was responsible for Api in our implementation,
I'd prefer to save it.

- the 'prefix' option related to common files is always constant
in our cmake function. The reason is Qt library files are
pre-generated and their names can not change from generation
to generation.
Note: users can change 'prefix' option manually from the command
interface.

- Set the modelNamePrefix empty for all models by default,
as discussed.
We don't want to use any Qt related signature for them, they are not
a part of Qt classes.

Task-number: QTBUG-142395
Change-Id: Ic8c043db9b747f61e1a19c1ae564ba86a079833a
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>
(cherry picked from commit 5dd7b2617bfa07011188b89971ae24b2487612e4)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>