Make ServerVariable a proper value type
* De-inline it and export out-of-line methods.
* Move the member variables into d_ptr, so that the class
  could be extended in a BC way.
* Add all 5 SMF + swap().
* Add (in)equality operators.
* Delete qHash() overloads.

A couple of remarks.

1. I chose to not add qHash() overloads, because for now I do not
see a value in using this class inside a QHash. I marked the
overloads as deleted, so that the users do not provide their own.
We can add the overloads later if needed.

2. We need to remember that the code in the common library is not
necessarily built as a part of Qt. Users can choose to build it
separately. As a result, we cannot use the
QT_DECLARE_QESDP_SPECIALIZATION_DTOR macro, and so we cannot have
the move SMFs and swap() declared inline.

Found in API review.

Task-number: QTBUG-142312
Change-Id: I8a3446f51b4f22ce39e776189a06c3290fa9141e
Reviewed-by: Rym Bouabid <rym.bouabid@qt.io>
Reviewed-by: Tatiana Borisova <tatiana.borisova@qt.io>
(cherry picked from commit 1e82771a3a6d03e3f9f75b7c586ec1ffb2d3d9e6)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>