Use a 'packageName' for CombinedModelsAndAPIs and export files
- use the 'packageName' (which is a cmake target) for naming
<*>Export.h, <*>CommonExports.h and <*>CombinedModelsAndAPIs.cpp
files on generation.
- use a 'packageNameUpperCase' instead of the 'prefix' for export
macros names.

The reason of changing is a user can set an empty prefix using
MODEL_NAME_PREFIX command's argument.
It will lead to the bad export naming, like: '_EXPORT'

At the same time the 'packageName' should always be set by the user,
so it is safer to use 'packageName' for export macros
and for CombinedModelsAndAPIs.

Pick-to: 6.11
Change-Id: I1c8ef50267d752e1347d9aeb1f36e6c1352f8504
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>