Make ServerConfiguration a proper value type
* De-inline it and export out-of-line methods.
* Move the member variables into d_ptr, so that the class could be
  extended in a BC way.
* Add all 5 SMF + swap().
* Add (in)equality operators.
* Delete qHash() overloads.
* Rename setDefaultValue() to setVariableValue() as a it is a
  better name for what this function does.
* Rewrite setVariableValue() implementation to avoid calling both
  contains() and operator[]. Use a non-const find() and operate
  on the returned iterator instead.

A couple of remarks.

1. I chose to not add qHash() overloads, because for now I do not
see a value in using this class inside a QHash. I marked the
overloads as deleted, so that the users do not provide their own.
We can add the overloads later if needed.

2. We need to remember that the code in the common library is not
necessarily built as a part of Qt. Users can choose to build it
separately. As a result, we cannot use the
QT_DECLARE_QESDP_SPECIALIZATION_DTOR macro, and so we cannot have
the move SMFs and swap() declared inline.

Found in API review.

Task-number: QTBUG-142312
Pick-to: 6.11
Change-Id: I5050f141a5fecdca6dd8aa17d7c39f2ba0929dcc
Reviewed-by: Tatiana Borisova <tatiana.borisova@qt.io>
Reviewed-by: Rym Bouabid <rym.bouabid@qt.io>