Rework BaseApi::setNetworkAccessResources()
We do not need to pass a pointer to QNetworkAccessManager, because it
can be easily accessed from QRestAccessManager, so drop the first
parameter.

Now when the function takes only one parameter we can rename it to
setRestAccessManager() to better align with what it is doing.

Found in API review.

Task-number: QTBUG-142312
Pick-to: 6.11
Change-Id: Idcf620811543fbbee772bf3ca9aa0af44c8be084
Reviewed-by: Tatiana Borisova <tatiana.borisova@qt.io>