Rework BaseApi::setNetworkAccessResources()
We do not need to pass a pointer to QNetworkAccessManager, because it
can be easily accessed from QRestAccessManager, so drop the first
parameter.

Now when the function takes only one parameter we can rename it to
setRestAccessManager() to better align with what it is doing.

Found in API review.

Task-number: QTBUG-142312
Change-Id: Idcf620811543fbbee772bf3ca9aa0af44c8be084
Reviewed-by: Tatiana Borisova <tatiana.borisova@qt.io>
(cherry picked from commit 5ddd7ef5690f149be2fb30a04f6aa5ba92a52845)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>