BaseApi: do not use std::shared_ptr for QNetworkRequestFactory
I could come up with only one reason to use a shared_ptr to store the
factory: someone might want to have a common factory and share it among
multiple API classes.

However, each BaseApi class has a couple of setters (namely, setTimeOut()
and setSslConfiguration()), that unconditionally change the state of the
shared factory.

This can be surprising for the other classes that use the same factory,
so I consider it a bad API.
This patch gets rid of the shared pointer and stores the object by value.
The setter is also updated to take the factory by value.
This is cheap since it's a typical Qt value type with a shared d_ptr.

This patch also adds a BaseApi::sslConfiguration() getter, because it is
now required in the PetApiTests::sslConfigurationTest().

Found in API review.

Task-number: QTBUG-142312
Change-Id: Ibee91c7d98165843c57fd81931e88bb1fed583f3
Reviewed-by: Tatiana Borisova <tatiana.borisova@qt.io>
(cherry picked from commit 189dd89cde8485be44fc5ec1a946e5f1a99e94c4)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>